+++ ambition_utils/rrule/tests/model_tests.py
@@ -1,11 +1,15 @@
 import datetime
+import fleming
 
+from fleming import fleming
 import pytz
-from dateutil import rrule
-from django_dynamic_fixture import G
+from dateutil import rrule, parser
 from django.test import TestCase
+from django_dynamic_fixture import G
 from freezegun import freeze_time
 
+from ambition_utils.rrule.constants import RecurrenceEnds
+from ambition_utils.rrule.forms import RecurrenceForm
 from ambition_utils.rrule.handler import OccurrenceHandler
 from ambition_utils.rrule.models import RRule
 from ambition_utils.rrule.tests.models import Program
@@ -15,6 +19,7 @@ class MockHandler(OccurrenceHandler):
     """
     Mack handler for handling an occurrence during testing
     """
+
     def handle(self, rrule):
         return True
 
@@ -445,6 +450,59 @@ class RRuleTest(TestCase):
         # Coverage for returning early
         self.assertEqual(rule.update_next_occurrence(), None)
 
+    def test_get_next_occurrence_dst(self):
+        """
+        What happens across DST changes?
+        """
+
+        # Create the params to create the rule
+        params = {
+            'freq': rrule.DAILY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2022, 10, 29),
+            'until': datetime.datetime(2022, 11, 1),
+            'byhour': 10,
+        }
+        
+        timezone = pytz.timezone('Europe/Kiev')
+        format = '%Y-%m-%d %H:%M'
+
+        # Create the rule
+        rule = RRule.objects.create(
+            rrule_params=params,
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler',
+            time_zone=timezone,
+        )
+
+        # Assert the initial values
+        # Europe/Kiev is UTC + 3 prior to 10/30/22. (10 Europe/Kiev meeting is 7 UTC.)
+        self.assertEqual(rule.last_occurrence, None)
+        self.assertEqual(rule.next_occurrence, datetime.datetime(2022, 10, 29, 7))
+        self.assertEqual(
+            fleming.convert_to_tz(rule.next_occurrence, timezone).strftime(format),
+            '2022-10-29 10:00'
+        )
+
+        # Notice next occurrence jumps to UTC + 2 to reflect change from DST on early hours of 10/30.
+        # Notice the converted date is still the expected 10am.
+        rule.update_next_occurrence()
+        self.assertEqual(rule.last_occurrence, datetime.datetime(2022, 10, 29, 7))
+        self.assertEqual(rule.next_occurrence, datetime.datetime(2022, 10, 30, 8))
+        self.assertEqual(
+            fleming.convert_to_tz(rule.next_occurrence, timezone).strftime(format),
+            '2022-10-30 10:00'
+        )
+
+        # Notice UTC + 2 is here to stay.
+        # Notice the converted date is still the expected 10am.
+        rule.update_next_occurrence()
+        self.assertEqual(rule.last_occurrence, datetime.datetime(2022, 10, 30, 8))
+        self.assertEqual(rule.next_occurrence, datetime.datetime(2022, 10, 31, 8))
+        self.assertEqual(
+            fleming.convert_to_tz(rule.next_occurrence, timezone).strftime(format),
+            '2022-10-31 10:00'
+        )
+
     def test_model_default_time_zone(self):
         params = {
             'freq': rrule.DAILY,
@@ -630,7 +688,7 @@ class RRuleTest(TestCase):
 
     def test_generate_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Assert generate_dates returns the same values as get_dates.
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -645,23 +703,122 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        
+        self.assertEqual(
+            rule.get_dates(),
+            rule.generate_dates()
+        )
+        
+    def test_get_dates(self):
+        """
+        Test a monthly 1st day of month rule to catch case of converting tz back using the get_dates method
+        """
+        params = {
+            'freq': rrule.MONTHLY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2016, 12, 31),
+            'bymonthday': 1,
+            'until': datetime.datetime(2017, 4, 30),
+        }
+
+        rule = RRule(
+            rrule_params=params,
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates()
 
         # Check a few dates
         self.assertEqual(len(next_dates), 4)
-        self.assertEqual(next_dates[0], datetime.datetime(2017, 2, 1, 3))
-        self.assertEqual(next_dates[1], datetime.datetime(2017, 3, 1, 3))
-        self.assertEqual(next_dates[2], datetime.datetime(2017, 4, 1, 2))
-        self.assertEqual(next_dates[3], datetime.datetime(2017, 5, 1, 2))
+
+        self.assertEqual(next_dates[0], datetime.datetime(2017, 1, 1, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2017, 2, 1, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2017, 3, 1, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2017, 4, 1, 4))  # DST change for US/Eastern
 
         # Run pre save again to make sure it doesn't mess up params
         rule.pre_save_hooks()
 
         # Get next dates to compare against
-        more_next_dates = rule.generate_dates()
+        more_next_dates = rule.get_dates()
         self.assertEqual(next_dates, more_next_dates)
+        
+    def test_get_dates_with_start_date(self):
+        """
+        Test a date generation with a start date.
+        """
+        params = {
+            'freq': rrule.MONTHLY,
+            'interval': 1,
+            'dtstart': datetime.datetime(2016, 12, 31),
+            'bymonthday': 1,
+        }
+
+        rule = RRule(
+            rrule_params=params,
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates(num_dates=10, start_date=datetime.datetime(2018, 1, 1))
+
+        # Check a few dates
+        self.assertEqual(len(next_dates), 10)
+        self.assertEqual(next_dates[0], datetime.datetime(2018, 1, 1, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2018, 2, 1, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2018, 3, 1, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2018, 4, 1, 4))  # DST change for US/Eastern
+        self.assertEqual(next_dates[-1], datetime.datetime(2018, 10, 1, 4))
+
+    def test_get_dates_num_dates_greater(self):
+        """
+        Test a date generation with a start date and end date that will yield fewer dates than num_dates.
+        Daily from 1/1 to 1/10 is 10. Request default of 20 but only after 1/5 which should yield 6 dates, inclusive.
+        """
+        rule = RRule(
+            rrule_params={
+                'freq': rrule.DAILY,
+                'interval': 1,
+                'dtstart': datetime.datetime(2017, 1, 1),
+                'until': datetime.datetime(2017, 1, 10),
+            },
+            time_zone=pytz.timezone('US/Eastern'),
+            occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
+        )
+        next_dates = rule.get_dates(
+            num_dates=20,
+            start_date=datetime.datetime(2017, 1, 5)
+        )
+
+        self.assertEqual(len(next_dates), 6)
+        self.assertEqual(next_dates[0], datetime.datetime(2017, 1, 5, 5))
+        self.assertEqual(next_dates[1], datetime.datetime(2017, 1, 6, 5))
+        self.assertEqual(next_dates[2], datetime.datetime(2017, 1, 7, 5))
+        self.assertEqual(next_dates[3], datetime.datetime(2017, 1, 8, 5))
+        self.assertEqual(next_dates[4], datetime.datetime(2017, 1, 9, 5))
+        self.assertEqual(next_dates[5], datetime.datetime(2017, 1, 10, 5))
 
     def test_generate_dates_from_params(self):
+        """
+        Assert generate_dates_from_params returns the same values as get_dates_from_params.
+        """
+        params = {
+            'rrule_params': {
+                'freq': rrule.MONTHLY,
+                'interval': 1,
+                'dtstart': datetime.datetime(2017, 1, 1, 22),
+                'bymonthday': -1,
+                'until': datetime.datetime(2017, 5, 1, 22),
+            },
+            'time_zone': pytz.timezone('US/Eastern'),
+            'num_dates': 3,
+        }
+
+        self.assertEqual(
+            RRule.get_dates_from_params(**params),
+            RRule.generate_dates_from_params(**params)
+        )
+
+    def test_get_dates_from_params(self):
         """
         Tests the class method wrapper
         """
@@ -678,19 +835,23 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates(num_dates=3)
+        next_dates = rule.get_dates(
+            num_dates=3, 
+            start_date=datetime.datetime(2017, 2, 2, 22)
+        )
 
-        next_dates_from_params = RRule.generate_dates_from_params(
+        next_dates_from_params = RRule.get_dates_from_params(
             rrule_params=params,
             time_zone=pytz.timezone('US/Eastern'),
             num_dates=3,
+            start_date=datetime.datetime(2017, 2, 2, 22)
         )
 
         self.assertEqual(next_dates, next_dates_from_params)
 
-    def test_model_different_time_zone_end_of_month_generate_dates(self):
+    def test_model_different_time_zone_end_of_month_get_dates(self):
         """
-        Test a monthly first day of month rule to catch case of converting tz back using the generate_dates method
+        Test a monthly first day of month rule to catch case of converting tz back using the get_dates method
         """
         params = {
             'freq': rrule.MONTHLY,
@@ -704,7 +865,7 @@ class RRuleTest(TestCase):
             time_zone=pytz.timezone('US/Eastern'),
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
         )
-        next_dates = rule.generate_dates()
+        next_dates = rule.get_dates()
         self.assertEqual(next_dates[0], datetime.datetime(2017, 2, 1, 3))
         self.assertEqual(next_dates[1], datetime.datetime(2017, 3, 1, 3))
         self.assertEqual(next_dates[2], datetime.datetime(2017, 4, 1, 2))
@@ -933,7 +1094,6 @@ class RRuleTest(TestCase):
 
         # Save the rrule with a current time before the first occurrence
         with freeze_time('1-1-2016'):
-
             # Change the start date to a future date
             rule.rrule_params['dtstart'] = datetime.datetime(2018, 1, 1)
             rule.refresh_next_occurrence()
@@ -1026,16 +1186,16 @@ class RRuleTest(TestCase):
         self.assertEqual(rule.rrule_params, clone.rrule_params)
 
         # Assert the generated dates are equal.
-        self.assertEqual(rule.generate_dates(num_dates=4), clone.generate_dates(num_dates=4))
+        self.assertEqual(rule.get_dates(num_dates=4), clone.get_dates(num_dates=4))
 
     @freeze_time('6-15-2022')
-    def test_clone_with_offset(self):
+    def test_weekly_clone_with_offset(self):
         # New object that starts next Wednesday
         # Weekly on MWF
         rule = RRule.objects.create(
             rrule_params={
                 'freq': rrule.WEEKLY,
-                'dtstart': datetime.datetime(2022, 6, 22),
+                'dtstart': datetime.datetime(2022, 6, 21),  # Tuesday
                 'byweekday': [0, 2, 4],
             },
             occurrence_handler_path='ambition_utils.rrule.tests.model_tests.MockHandler'
@@ -1060,36 +1220,206 @@ class RRuleTest(TestCase):
 
         # Assert the generated dates are as expected.
         self.assertEqual(
-            rule.generate_dates(num_dates=4),
+            rule.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 22),  # Wednesday
                 datetime.datetime(2022, 6, 24),  # Friday
                 datetime.datetime(2022, 6, 27),  # Monday
                 datetime.datetime(2022, 6, 29),  # Wednesday
-                datetime.datetime(2022, 7, 1),   # Friday
             ]
         )
 
         # Two days after each date in the regular series.
         self.assertEqual(
-            future_clone.generate_dates(num_dates=4),
+            future_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 24),  # Friday
                 datetime.datetime(2022, 6, 26),  # Sunday
                 datetime.datetime(2022, 6, 29),  # Wednesday
                 datetime.datetime(2022, 7, 1),   # Friday
-                datetime.datetime(2022, 7, 3),   # Sunday
             ]
         )
 
         # Two days before each date in the regular series.
         self.assertEqual(
-            past_clone.generate_dates(num_dates=4),
+            past_clone.get_dates(num_dates=4),
             [
                 datetime.datetime(2022, 6, 20),  # Monday
                 datetime.datetime(2022, 6, 22),  # Wednesday
                 datetime.datetime(2022, 6, 25),  # Saturday
                 datetime.datetime(2022, 6, 27),  # Monday
-                datetime.datetime(2022, 6, 29),  # Wednesday
+            ]
+        )
+
+    @freeze_time('6-1-2022')
+    def test_monthly_clone_with_offset(self):
+        """
+        Assert than an object with bynweekday data can be cloned with an offset.
+        """
+
+        # Second Monday from end of month every other month
+        # Starts today, 6/1. First occurrence is 6/20. (Last Monday is 27th.)
+        data = {
+            'freq': rrule.MONTHLY,
+            'interval': 2,
+            'dtstart': '6/1/2022',
+            'byhour': '0',
+            'time_zone': 'UTC',
+            'ends': RecurrenceEnds.NEVER,
+            'repeat_by': 'DAY_OF_THE_WEEK_END',
+            'bynweekday': '[[0, -2]]'
+        }
+
+        # Form is used to flex the bynweeday to byweekday + bysetpos conversion that occurs in its save().
+        form = RecurrenceForm(data=data)
+        self.assertTrue(form.is_valid())
+        rule = form.save()
+
+        # Create a clones of the object with a start date of 2 days into the future and 2 days into the past.
+        future_clone = rule.clone_with_day_offset(2)
+        past_clone = rule.clone_with_day_offset(-2)
+
+        # Assert that the rule created here is unchanged but the clones reflect their offsets.
+        format = '%Y-%m-%d'
+        self.assertEqual(rule.next_occurrence.strftime(format), '2022-06-20')
+        self.assertEqual(future_clone.next_occurrence.strftime(format), '2022-06-22')
+        self.assertEqual(past_clone.next_occurrence.strftime(format), '2022-06-18')
+
+    @freeze_time('6-15-2022')
+    def test_clone_with_day_offset_with_finish_date(self):
+        """
+        Assert that the 'until' param is offset correctly
+        """
+
+        rule = RRule.objects.create(
+            rrule_params={
+                'freq': rrule.DAILY,
+                'dtstart': datetime.datetime(2022, 10, 15),
+                'until': datetime.datetime(2022, 10, 17)
+            }
+        )
+
+        future_clone = rule.clone_with_day_offset(1)
+        past_clone = rule.clone_with_day_offset(-1)
+
+        # Assert the updated until values are correct
+        self.assertEqual(
+            parser.parse(rule.rrule_params['until']) + datetime.timedelta(days=1),
+            parser.parse(future_clone.rrule_params['until'])
+        )
+        self.assertEqual(
+            parser.parse(rule.rrule_params['until']) - datetime.timedelta(days=1),
+            parser.parse(past_clone.rrule_params['until'])
+        )
+
+        # Assert the generated dates are as expected for the original.
+        self.assertEqual(
+            rule.get_dates(),
+            [
+                datetime.datetime(2022, 10, 15),
+                datetime.datetime(2022, 10, 16),
+                datetime.datetime(2022, 10, 17)
+            ]
+        )
+
+        # One day after each date in the regular series.
+        self.assertEqual(
+            future_clone.get_dates(),
+            [
+                datetime.datetime(2022, 10, 16),
+                datetime.datetime(2022, 10, 17),
+                datetime.datetime(2022, 10, 18)
+            ]
+        )
+
+        # One day before each date in the regular series.
+        self.assertEqual(
+            past_clone.get_dates(),
+            [
+                datetime.datetime(2022, 10, 14),
+                datetime.datetime(2022, 10, 15),
+                datetime.datetime(2022, 10, 16)
+            ]
+        )
+
+    @freeze_time('10-31-2022')
+    def test_clone_with_future_days_across_dst(self):
+        """
+        Assert that the resulting recurrence next occurrence reflects *its* timezone offset. 
+        """
+
+        rule = RRule.objects.create(
+            rrule_params={
+                'freq': rrule.DAILY,
+                'dtstart': datetime.datetime(2022, 10, 29, 10),
+                'until': datetime.datetime(2022, 11, 1, 10),
+            },
+            time_zone=pytz.timezone('Europe/Kiev')
+        )
+       
+        # Europe/Kiev goes from UTC+3 to UTC+2 in the early hours of 10/30.
+        self.assertEqual(
+            rule.generate_dates(),
+            [
+                datetime.datetime(2022, 10, 29, 7),
+                datetime.datetime(2022, 10, 30, 8),
+                datetime.datetime(2022, 10, 31, 8),
+                datetime.datetime(2022, 11, 1, 8),
+            ]
+        )
+
+        # Europe/Kiev goes to standard time (UTC+2) in the early hours of 1/30.
+        # This offset should result in a datetime (UTC+3).
+        past_clone = rule.clone_with_day_offset(-1)
+
+        # One day before each date in the regular series.
+        self.assertEqual(
+            past_clone.generate_dates(),
+            [
+                datetime.datetime(2022, 10, 28, 7),
+                datetime.datetime(2022, 10, 29, 7),
+                datetime.datetime(2022, 10, 30, 8),
+                datetime.datetime(2022, 10, 31, 8),
+            ]
+        )
+
+    @freeze_time('10-31-2022')
+    def test_clone_with_offset_before_dst(self):
+        """
+        Assert that a clone, with a negative offset that puts it in DST, results in dates that respect the
+        DST transition once it gets there.
+        Europe/Kiev goes from UTC+3 to UTC+2 in the early hours of 10/30.
+        """
+
+        # Starting in standard time
+        rule = RRule.objects.create(
+            rrule_params={
+                'freq': rrule.DAILY,
+                'dtstart': datetime.datetime(2022, 10, 31, 10),
+                'until': datetime.datetime(2022, 11, 3, 10),
+            },
+            time_zone=pytz.timezone('Europe/Kiev')
+        )
+
+        # Times are UTC+2 because it is after the 10/30 transition.
+        self.assertEqual(
+            rule.generate_dates(),
+            [
+                datetime.datetime(2022, 10, 31, 8),
+                datetime.datetime(2022, 11, 1, 8),
+                datetime.datetime(2022, 11, 2, 8),
+                datetime.datetime(2022, 11, 3, 8),
+            ]
+        )
+
+        # Clone prior to the DST switch and ensure the times transition between 10/30 & 10/31.
+        past_clone = rule.clone_with_day_offset(-3)
+        self.assertEqual(
+            past_clone.generate_dates(),
+            [
+                datetime.datetime(2022, 10, 28, 7),
+                datetime.datetime(2022, 10, 29, 7),
+                datetime.datetime(2022, 10, 30, 8),
+                datetime.datetime(2022, 10, 31, 8),
             ]
         )