+++ ambition_utils/docs/release_notes.rst
@@ -1,3 +1,4 @@
+
 Release Notes
 =============
 
@@ -8,6 +9,27 @@ Release Notes
 * Add support for Django 3.0, 3.1, 3.2, 4.0, 4.1
 * Drop support for django 2.0, 2.1
 * Switch to github actions
+
+2.6.0
+-----
+* Added get_dates(start_date=) to calculate dates after start_date. Will replace get_dates.
+* Added get_dates_from_params(start_date=) to calculate dates after start_date from rrule_params. Will replace generate_dates_from_params.
+
+2.5.3
+-----
+* Fixed cloning when until param is present
+
+2.5.2
+-----
+* Fixed cloning with bynweekday data
+
+2.5.1
+-----
+* Updated the get_time_zone() helper method with pytz.all_timezones
+
+2.5.0
+-----
+* Added support for byminute
 
 2.4.0
 -----